+++ src/ts/components/hugeplaybacktogglebutton.ts
@@ -124,6 +124,14 @@ export class HugePlaybackToggleButton extends PlaybackToggleButton {
     player.addEventHandler(player.EVENT.ON_CAST_START, castInitializationHandler);
     player.addEventHandler(player.EVENT.ON_CAST_STARTED, castInitializationHandler);
     player.addEventHandler(player.EVENT.ON_CAST_STOPPED, castInitializationHandler);
+
+    // Hide the play button animation when the UI is loaded (it should only be animated on state changes)
+    this.setTransitionAnimationsEnabled(false);
+
+    // Enable the transition animations when the play state changes for the first time
+    this.onToggle.subscribeOnce(() => {
+      this.setTransitionAnimationsEnabled(true);
+    });
   }
 
   protected toDomElement(): DOM {
@@ -138,5 +146,20 @@ export class HugePlaybackToggleButton extends PlaybackToggleButton {
     }));
 
     return buttonElement;
+  }
+
+  /**
+   * Enables or disables the play state transition animations of the play button image. Can be used to suppress
+   * animations.
+   * @param {boolean} enabled true to enable the animations (default), false to disable them
+   */
+  protected setTransitionAnimationsEnabled(enabled: boolean): void {
+    const noTransitionAnimationsClass = this.prefixCss('no-transition-animations');
+
+    if (enabled) {
+      this.getDomElement().removeClass(noTransitionAnimationsClass);
+    } else if (!this.getDomElement().hasClass(noTransitionAnimationsClass)) {
+      this.getDomElement().addClass(noTransitionAnimationsClass);
+    }
   }
 }